pyproject.toml
[project]
name = "sec10q-rag"
version = "0.1.0"
description = "RAG pipeline over SEC 10-Q filings using SentenceTransformers + FAISS."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0",
  "beautifulsoup4>=4.12.0",
  "lxml>=5.0.0",
  "numpy>=1.24.0",
  "faiss-cpu>=1.8.0",
  "sentence-transformers>=2.6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
